경북대 BE_배준호 6주차 과제(1단계) (#169)
* feat : step 0 기본 코드 준비

* build : build 파일 설정

* refactor : 회원 API 명세 수정

* refactor : 과제 힌트에 따라 사양 변경

기존에 field와 sort 두 개의 필드로 받던 값을 sort 하나에 배열 형식으로 받는 방식으로 변경

* refactor : 과제 힌트에 따라 사양 변경

기존에 field와 sort 두 개의 필드로 받던 값을 sort 하나에 배열 형식으로 받는 방식으로 변경

* feat : 주문 조회 기능 추가 ( 페이징 적용 )

* refactor : 페이징 적용하지 않도록 변경

* refactor : 카테고리, 옵션 페이지네이션 지우기

* refactor : 페이지네이션 없는 버전의 테스트 구현

* refactor : 관리자 페이지에서 로그인 및 회원가입 구현

* refactor : 오타 수정 및 관리자 페이지수정 사항 반영

* refactor : 요구 사항을 보고 추측을 통해 새로운 필드 추가

* fix : 페이지 방식이 아닌 한 번에 불러오는 방식으로 수정

* refactor : 생성자에 새로운 필드 추가

* refactor : 수정사항 반영 및 파일 형식 변환

* docs : 요구사항 추가

* chore : 사용하지 않는 exception 명시 삭제

* refactor : 유저의 userId 필드 삭제

* refactor : 상품 조회에 categoryId 조건 포함

* refactor : 카카오 로그인 POST로 수정

* refactor : 사용하지 않는 sql 전처리 삭제

* refactor : 유저 도메인 리펙터링

협의를 통해 정한 API 사양에 맞추기

* refactor : 위시리스트 도메인 리펙터링

협의를 통해 정한 API 사양에 맞추기

* refactor : 상품 도메인 리펙터링

협의를 통해 정한 API 사양에 맞추기

* refactor : 카테고리 도메인 리펙터링

협의를 통해 정한 API 사양에 맞추기

* refactor : 옵션 도메인 리펙터링

협의를 통해 정한 API 사양에 맞추기

* refactor : 주문 도메인 리펙터링

협의를 통해 정한 API 사양에 맞추기

* feat : CORS 에러를 예방하기 위한 헤더 설정을 글로벌하게 설정

* feat : 팀원과 합의하에 만든 API 문서

* refactor : 일대다 양방향 매핑으로 변경

사양을 수정하다보니 이렇게 하는 것이 효과적일 것이라고 결정되었다

* refactor : 연관관계 설정 변경에 따른 리펙터링

* refactor : 연관관계 설정 변경에 따른 리펙터링

* refactor : 사용하지 않는 파일 삭제

* refactor : 전처리를 통한 카테고리 등록 재구현

* refactor : 연관관계 변경에 따른 리펙터링

* refactor : 카테고리 도메인 필드 추가에 따른 리펙터링

* refactor : 사라진 유저 아이디 대신 이메일로 표기

* refactor : 연관관계 변경에 따른 리펙터링

* refactor : 관리자 페이지를 메인 페이지로 변경

* refactor : 유저 정보 조회에 토큰이 필요하도록 변경

* refactor : 더 이상 사용하지 않는 유저 아이디 필드 삭제

* 미구현 사항

* refactor : 프론트 조건에 맞추기 위한 제약 조건 수정

* fix : 리펙터링 사항 반영

* refactor : 카카오 로그인 처리 방식 결정

* refactor : 접근 허용 오리진 모든 오리진으로 변경

* refactor : 도메인 전처리 추가

* fix : 모든 오리진을 제대로 허용하도록 변경

* refactor : 토큰 전송 방법 결정 및 반영

* fix : 에러 수정

* 자동 빌드 및 실행 스크립트 작성

---------

+++ src/main/java/gift/domain/Order.java
@@ -14,7 +14,7 @@ public class Order extends BaseEntity {
     @Column(nullable = false, name="option_id")
     private Long optionId;
     @Column(nullable = false, name="quantity")
-    private Long quantity;
+    private int quantity;
     @Column(nullable = false, name="message")
     private String message;
     @CreatedDate
@@ -26,7 +26,7 @@ public class Order extends BaseEntity {
         super();
     }
 
-    public Order(Long optionId, Long quantity, String message) {
+    public Order(Long optionId, int quantity, String message) {
         this.optionId = optionId;
         this.quantity = quantity;
         this.message = message;
@@ -36,7 +36,7 @@ public class Order extends BaseEntity {
         return optionId;
     }
 
-    public Long getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
